# data-api/openapi.yaml
openapi: 3.1.0
info:
  title: Local-first RAG API
  version: 1.0.0-draft.1
  description: Account scope is always derived from the session cookie; opaque IDs are server generated.
servers: [{url: /api/v1}]
paths:
  /auth/register:
    post:
      summary: Create account and session
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RegisterRequest'}}}}
      responses: {'201': {description: Authenticated account, content: {application/json: {schema: {$ref: '#/components/schemas/Account'}}}}, '400': {$ref: '#/components/responses/ValidationError'}, '409': {$ref: '#/components/responses/Conflict'}}
  /auth/login:
    post:
      summary: Start session
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}}
      responses: {'200': {description: Authenticated account, content: {application/json: {schema: {$ref: '#/components/schemas/Account'}}}}, '401': {$ref: '#/components/responses/Unauthenticated'}}
  /auth/logout:
    post: {security: [{sessionCookie: []}], responses: {'204': {description: Session revoked}, '401': {$ref: '#/components/responses/Unauthenticated'}}}
  /auth/me:
    get: {security: [{sessionCookie: []}], responses: {'200': {description: Account, content: {application/json: {schema: {$ref: '#/components/schemas/Account'}}}}, '401': {$ref: '#/components/responses/Unauthenticated'}}}
  /documents:
    get:
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Cursor page, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentPage'}}}}, '401': {$ref: '#/components/responses/Unauthenticated'}}
    post:
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary}}}}}}
      responses: {'201': {description: Uploaded, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}, '400': {$ref: '#/components/responses/ValidationError'}, '413': {$ref: '#/components/responses/Oversize'}, '415': {$ref: '#/components/responses/UnsupportedType'}, '409': {$ref: '#/components/responses/Conflict'}}
  /documents/{documentId}:
    parameters: [{$ref: '#/components/parameters/DocumentId'}]
    get: {security: [{sessionCookie: []}], responses: {'200': {description: Document, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}, '404': {$ref: '#/components/responses/NotFound'}}}
    delete: {security: [{sessionCookie: []}], responses: {'204': {description: Deleted; repeat is safe}, '404': {$ref: '#/components/responses/NotFound'}}}
  /documents/{documentId}/chunks/{chunkId}:
    parameters: [{$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/ChunkId'}]
    get: {security: [{sessionCookie: []}], responses: {'200': {description: Scoped citation excerpt, content: {application/json: {schema: {$ref: '#/components/schemas/Chunk'}}}}, '404': {$ref: '#/components/responses/NotFound'}}}
  /conversations:
    get:
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Cursor page, content: {application/json: {schema: {$ref: '#/components/schemas/ConversationPage'}}}}}
    post:
      security: [{sessionCookie: []}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [title], properties: {title: {type: string, minLength: 1, maxLength: 200}}}}}}
      responses: {'201': {description: Conversation, content: {application/json: {schema: {$ref: '#/components/schemas/Conversation'}}}}, '400': {$ref: '#/components/responses/ValidationError'}}
  /conversations/{conversationId}:
    parameters: [{$ref: '#/components/parameters/ConversationId'}]
    get: {security: [{sessionCookie: []}], responses: {'200': {description: Conversation, content: {application/json: {schema: {$ref: '#/components/schemas/Conversation'}}}}, '404': {$ref: '#/components/responses/NotFound'}}}
    patch:
      security: [{sessionCookie: []}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [title], properties: {title: {type: string, minLength: 1, maxLength: 200}}}}}}
      responses: {'200': {description: Conversation, content: {application/json: {schema: {$ref: '#/components/schemas/Conversation'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
  /conversations/{conversationId}/messages:
    parameters: [{$ref: '#/components/parameters/ConversationId'}]
    get:
      security: [{sessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Cursor page, content: {application/json: {schema: {$ref: '#/components/schemas/MessagePage'}}}}, '404': {$ref: '#/components/responses/NotFound'}}
    post:
      security: [{sessionCookie: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AskRequest'}}}}
      responses: {'200': {description: Prior idempotent result, content: {application/json: {schema: {$ref: '#/components/schemas/AskResponse'}}}}, '201': {description: Answer attempt completed, content: {application/json: {schema: {$ref: '#/components/schemas/AskResponse'}}}}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '429': {$ref: '#/components/responses/Busy'}, '502': {$ref: '#/components/responses/ModelFailure'}, '504': {$ref: '#/components/responses/ModelFailure'}}
components:
  securitySchemes: {sessionCookie: {type: apiKey, in: cookie, name: session}}
  parameters:
    DocumentId: {name: documentId, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueId'}}
    ChunkId: {name: chunkId, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueId'}}
    ConversationId: {name: conversationId, in: path, required: true, schema: {$ref: '#/components/schemas/OpaqueId'}}
    Cursor: {name: cursor, in: query, schema: {type: string, minLength: 16}, description: Opaque cursor from prior page}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
  responses:
    Unauthenticated: {description: No valid session, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    NotFound: {description: Missing or foreign resource (intentionally indistinguishable), content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: {code: not_found, message: Resource not found, requestId: req_opaque}}}}}
    ValidationError: {description: Invalid request, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Conflict: {description: State or idempotency conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Oversize: {description: File exceeds 25 MiB, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    UnsupportedType: {description: Unsupported or invalid file, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Busy: {description: Answer capacity exhausted, headers: {Retry-After: {schema: {type: integer}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ModelFailure: {description: Configured generation endpoint failed, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    OpaqueId: {type: string, pattern: '^[A-Za-z0-9_-]{32,64}$'}
    Error: {type: object, required: [error], properties: {error: {type: object, required: [code, message, requestId], properties: {code: {type: string, enum: [validation_error, unauthenticated, not_found, state_conflict, idempotency_conflict, file_too_large, unsupported_media_type, capacity_exhausted, model_unavailable, model_timeout]}, message: {type: string}, requestId: {type: string}}}}}
    Account: {type: object, required: [id], properties: {id: {$ref: '#/components/schemas/OpaqueId'}}}
    RegisterRequest: {type: object, required: [email, password], properties: {email: {type: string, format: email}, password: {type: string, minLength: 12, maxLength: 1024}}}
    LoginRequest: {type: object, required: [email, password], properties: {email: {type: string, format: email}, password: {type: string}}}
    Document: {type: object, required: [id, originalFilename, byteSize, status, createdAt], properties: {id: {$ref: '#/components/schemas/OpaqueId'}, originalFilename: {type: string}, byteSize: {type: integer}, status: {type: string, enum: [uploaded, processing, ready, failed]}, failureCode: {type: [string, 'null']}, createdAt: {type: string, format: date-time}}}
    Chunk: {type: object, required: [id, documentId, location, text], properties: {id: {$ref: '#/components/schemas/OpaqueId'}, documentId: {$ref: '#/components/schemas/OpaqueId'}, location: {type: object, required: [kind, start, end], properties: {kind: {type: string, enum: [page, line]}, start: {type: integer}, end: {type: integer}}}, text: {type: string}}}
    Conversation: {type: object, required: [id, title, createdAt, updatedAt], properties: {id: {$ref: '#/components/schemas/OpaqueId'}, title: {type: string}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}}}
    Message: {type: object, required: [id, role, content, status, createdAt], properties: {id: {$ref: '#/components/schemas/OpaqueId'}, role: {type: string, enum: [user, assistant]}, content: {type: string}, status: {type: string, enum: [completed, failed]}, createdAt: {type: string, format: date-time}}}
    Citation: {type: object, required: [documentId, documentName, chunkId, location], properties: {documentId: {$ref: '#/components/schemas/OpaqueId'}, documentName: {type: string}, chunkId: {$ref: '#/components/schemas/OpaqueId'}, location: {type: object, properties: {kind: {type: string, enum: [page, line]}, start: {type: integer}, end: {type: integer}}, required: [kind, start, end]}}}
    AskRequest: {type: object, required: [text, clientRequestId], properties: {text: {type: string, minLength: 1, maxLength: 12000}, clientRequestId: {type: string, minLength: 16, maxLength: 128}}}
    AskResponse: {type: object, required: [message, answer], properties: {message: {$ref: '#/components/schemas/Message'}, answer: {oneOf: [{type: object, required: [kind, text, citations], properties: {kind: {const: answered}, text: {type: string}, citations: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Citation'}}}}, {type: object, required: [kind, text, citations], properties: {kind: {enum: [no_documents, no_relevant_context]}, text: {type: string}, citations: {type: array, maxItems: 0}}}]}}}
    DocumentPage: {type: object, required: [items, nextCursor], properties: {items: {type: array, items: {$ref: '#/components/schemas/Document'}}, nextCursor: {type: [string, 'null']}}}
    ConversationPage: {type: object, required: [items, nextCursor], properties: {items: {type: array, items: {$ref: '#/components/schemas/Conversation'}}, nextCursor: {type: [string, 'null']}}}
    MessagePage: {type: object, required: [items, nextCursor], properties: {items: {type: array, items: {$ref: '#/components/schemas/Message'}}, nextCursor: {type: [string, 'null']}}}